Moved clamp fn to header file
