+++ include/fns.h
@@ -1,0 +1,16 @@
+#ifndef FNS_H
+#define FNS_H
+
+// Simple clamping function
+template <typename T>
+T clamp(T value, T minVal, T maxVal) {
+  if (value < minVal) {
+    return minVal;
+  } else if (value > maxVal) {
+    return maxVal;
+  } else {
+    return value;
+  }
+}
+
+#endif
